Refactor: move profile models to data.models
