+++ app/src/main/java/fr/free/nrw/commons/mwapi/OkHttpJsonApiClient.java
@@ -7,17 +7,16 @@ import android.text.TextUtils;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.google.gson.Gson;
-import fr.free.nrw.commons.auth.SessionManager;
 import fr.free.nrw.commons.campaigns.CampaignResponseDTO;
 import fr.free.nrw.commons.explore.depictions.DepictsClient;
 import fr.free.nrw.commons.location.LatLng;
 import fr.free.nrw.commons.nearby.Place;
 import fr.free.nrw.commons.nearby.model.NearbyResponse;
 import fr.free.nrw.commons.nearby.model.NearbyResultItem;
-import fr.free.nrw.commons.profile.achievements.FeaturedImages;
+import fr.free.nrw.commons.data.models.profile.FeaturedImages;
 import fr.free.nrw.commons.profile.achievements.FeedbackResponse;
-import fr.free.nrw.commons.profile.leaderboard.LeaderboardResponse;
-import fr.free.nrw.commons.profile.leaderboard.UpdateAvatarResponse;
+import fr.free.nrw.commons.data.models.profile.LeaderboardResponse;
+import fr.free.nrw.commons.data.models.profile.UpdateAvatarResponse;
 import fr.free.nrw.commons.upload.FileUtils;
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem;
 import fr.free.nrw.commons.utils.ConfigUtils;
