+++ app/src/test/kotlin/fr/free/nrw/commons/delete/ReasonBuilderTest.kt
@@ -6,8 +6,8 @@ import fr.free.nrw.commons.Media
 import fr.free.nrw.commons.auth.SessionManager
 import fr.free.nrw.commons.mwapi.OkHttpJsonApiClient
 import fr.free.nrw.commons.profile.achievements.FeedbackResponse
-import fr.free.nrw.commons.profile.leaderboard.LeaderboardResponse
-import fr.free.nrw.commons.profile.leaderboard.UpdateAvatarResponse
+import fr.free.nrw.commons.data.models.profile.LeaderboardResponse
+import fr.free.nrw.commons.data.models.profile.UpdateAvatarResponse
 import fr.free.nrw.commons.utils.ViewUtilWrapper
 import io.reactivex.Observable
 import io.reactivex.Single
