+++ app/src/test/kotlin/fr/free/nrw/commons/leaderboard/LeaderboardApiTest.java
@@ -1,7 +1,7 @@
 package fr.free.nrw.commons.leaderboard;
 
 import com.google.gson.Gson;
-import fr.free.nrw.commons.profile.leaderboard.LeaderboardResponse;
+import fr.free.nrw.commons.data.models.profile.LeaderboardResponse;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
